requirements.txt
fastapi==0.88.0
bcrypt==4.0.1
pyjwt==2.6.0
python-dotenv==0.21.0
python-multipart==0.0.5
git+https://github.com/SleepyDude/deta-async